[WDA-2] Pulled out the http data fetching into its own class for better encapsulation.
Made a DataFetcher class to grab data from a URL. Made it its own class primarily for two reasons: a) Easier to test against a multitude of urls, response types, data, etc. and b) so that I could make it generic so that it can fetch the json weather data and icons, etc.